Implement NSG rule management for CCM loadbalancers and add respective e2e test cases

+++ pkg/cloudprovider/providers/oci/load_balancer.go
@@ -17,6 +17,7 @@ package oci
 import (
 	"context"
 	"fmt"
+	"reflect"
 	"strconv"
 	"sync"
 	"time"
@@ -61,6 +62,10 @@ const DefaultLoadBalancerBEProtocol = "TCP"
 // balancer listeners created by the CCM.
 const DefaultNetworkLoadBalancerListenerProtocol = "TCP"
 
+// MaxNsgPerVnic is the maximum number of NSGs that can be attached to a vnic
+// https://docs.oracle.com/en-us/iaas/Content/General/Concepts/servicelimits.htm#nsg_limits
+const MaxNsgPerVnic = 5
+
 const (
 	// Fallback value if annotation on service is not set
 	lbDefaultShape = "100Mbps"
@@ -82,8 +87,8 @@ const (
 	serviceAccountTokenExpiry = 21600 // 6 Hours
 )
 
-// Protects Security Lists against update by multiple LBs in parallel
-var securityListMutex sync.Mutex
+// Protects security rule addition against update by multiple LBs in parallel
+var updateRulesMutex sync.Mutex
 
 // CloudLoadBalancerProvider is an implementation of the cloud-provider struct
 type CloudLoadBalancerProvider struct {
@@ -252,9 +257,9 @@ func getSubnetsForNodes(ctx context.Context, nodes []*v1.Node, client client.Int
 			return nil, errors.Errorf(".spec.providerID was not present on node %q", node.Name)
 		}
 
-		id, err := MapProviderIDToInstanceID(node.Spec.ProviderID)
+		id, err := MapProviderIDToResourceID(node.Spec.ProviderID)
 		if err != nil {
-			return nil, errors.Wrap(err, "MapProviderIDToInstanceID")
+			return nil, errors.Wrap(err, "MapProviderIDToResourceID")
 		}
 
 		compartmentID, ok := node.Annotations[CompartmentIDAnnotation]
@@ -501,6 +506,7 @@ func (cp *CloudProvider) EnsureLoadBalancer(ctx context.Context, clusterName str
 
 	var errorType string
 	var lbMetricDimension string
+	var nsgMetricDimension string
 
 	lbProvider, err := cp.getLoadBalancerProvider(ctx, service)
 	if err != nil {
@@ -596,6 +602,100 @@ func (cp *CloudProvider) EnsureLoadBalancer(ctx context.Context, clusterName str
 		metrics.SendMetricData(cp.metricPusher, getMetric(loadBalancerType, Update), time.Since(startTime).Seconds(), dimensionsMap)
 
 		return nil, err
+	}
+
+	if requiresNsgManagement(service) {
+		// Fetch existing frontend NSG and use it to manage rules
+		frontendNsgId := ""
+		backendNsgs := spec.ManagedNetworkSecurityGroup.backendNsgId
+
+		if lb != nil && lb.Id != nil && lb.NetworkSecurityGroupIds != nil {
+			nsgs := lb.NetworkSecurityGroupIds
+			for _, id := range nsgs {
+				frontendNsgId, _, err = cp.getFrontendNsg(ctx, logger, id, fmt.Sprintf("%s", service.UID))
+				if err != nil {
+					errorType = util.GetError(err)
+					nsgMetricDimension = util.GetMetricDimensionForComponent(errorType, util.NSGType)
+					dimensionsMap[metrics.ComponentDimension] = nsgMetricDimension
+					metrics.SendMetricData(cp.metricPusher, getMetric(util.NSGType, Get), time.Since(startTime).Seconds(), dimensionsMap)
+				}
+				if frontendNsgId != "" {
+					if !contains(spec.NetworkSecurityGroupIds, frontendNsgId) {
+						spec.NetworkSecurityGroupIds = append(spec.NetworkSecurityGroupIds, frontendNsgId)
+					}
+					logger.With("loadBalancerID", *lb.Id).Infof("using existing frontendNsg %s", frontendNsgId)
+					break
+				}
+			}
+		}
+		logger = logger.With("serviceId", service.UID)
+		// Create the NSG and add it to the LbSpec
+		if frontendNsgId == "" {
+			if len(spec.NetworkSecurityGroupIds) >= MaxNsgPerVnic {
+				return nil, fmt.Errorf("invalid number of Network Security Groups (Max: 5) including CCM managed nsg")
+			}
+			resp, err := cp.client.Networking().CreateNetworkSecurityGroup(ctx, cp.config.CompartmentID, cp.config.VCNID, generateNsgName(service), fmt.Sprintf("%s", service.UID))
+			if err != nil {
+				logger.With(zap.Error(err)).Error("Failed to create nsg")
+				errorType = util.GetError(err)
+				nsgMetricDimension = util.GetMetricDimensionForComponent(errorType, util.NSGType)
+				dimensionsMap[metrics.ComponentDimension] = nsgMetricDimension
+				dimensionsMap[metrics.ResourceOCIDDimension] = *resp.Id
+				metrics.SendMetricData(cp.metricPusher, getMetric(util.NSGType, Create), time.Since(startTime).Seconds(), dimensionsMap)
+				return nil, err
+			}
+			frontendNsgId = *resp.Id
+			spec.NetworkSecurityGroupIds = append(spec.NetworkSecurityGroupIds, frontendNsgId)
+			logger.With("frontendNsgId", *resp.Id).
+				Info("Successfully created nsg")
+			nsgMetricDimension = util.GetMetricDimensionForComponent(util.Success, util.NSGType)
+			dimensionsMap[metrics.ComponentDimension] = nsgMetricDimension
+			dimensionsMap[metrics.ResourceOCIDDimension] = *resp.Id
+			metrics.SendMetricData(cp.metricPusher, getMetric(util.NSGType, Create), time.Since(startTime).Seconds(), dimensionsMap)
+
+		}
+		if len(backendNsgs) > 0 {
+			for _, nsg := range backendNsgs {
+				resp, etag, err := cp.client.Networking().GetNetworkSecurityGroup(ctx, nsg)
+				if err != nil {
+					logger.With(zap.Error(err)).Error("Failed to get nsg")
+					errorType = util.GetError(err)
+					nsgMetricDimension = util.GetMetricDimensionForComponent(errorType, util.NSGType)
+					dimensionsMap[metrics.ComponentDimension] = nsgMetricDimension
+					dimensionsMap[metrics.ResourceOCIDDimension] = *resp.Id
+					metrics.SendMetricData(cp.metricPusher, getMetric(util.NSGType, Get), time.Since(startTime).Seconds(), dimensionsMap)
+					return nil, err
+				}
+				freeformTags := resp.FreeformTags
+				if _, ok := freeformTags["ManagedBy"]; !ok {
+					if etag != nil {
+						freeformTags["ManagedBy"] = "OKE-CCM"
+						response, err := cp.client.Networking().UpdateNetworkSecurityGroup(ctx, nsg, *etag, freeformTags)
+						if err != nil {
+							logger.With(zap.Error(err)).Error("Failed to get nsg")
+							errorType = util.GetError(err)
+							nsgMetricDimension = util.GetMetricDimensionForComponent(errorType, util.NSGType)
+							dimensionsMap[metrics.ComponentDimension] = nsgMetricDimension
+							dimensionsMap[metrics.ResourceOCIDDimension] = *response.Id
+							metrics.SendMetricData(cp.metricPusher, getMetric(util.NSGType, Update), time.Since(startTime).Seconds(), dimensionsMap)
+							return nil, err
+						}
+					}
+				}
+			}
+		}
+		serviceComponents := serviceComponents{
+			frontendNsgOcid:  frontendNsgId,
+			backendNsgOcids:  backendNsgs,
+			ports:            spec.Ports,
+			sourceCIDRs:      spec.SourceCIDRs,
+			isPreserveSource: *spec.IsPreserveSource,
+			serviceUid:       fmt.Sprintf("service-uid-%s", service.UID),
+		}
+		logger.Infof("(requiresNSGmanagement) Service Components %#v", serviceComponents)
+		if err = cp.reconcileSecurityGroup(ctx, serviceComponents); err != nil {
+			return nil, err
+		}
 	}
 
 	if !exists {
@@ -871,9 +971,8 @@ func (clb *CloudLoadBalancerProvider) updateLoadBalancer(ctx context.Context, lb
 }
 
 func updateSecurityListsInCriticalSection(ctx context.Context, spec *LBSpec, lbSubnets, nodeSubnets []*core.Subnet) (err error) {
-	securityListMutex.Lock()
-	defer securityListMutex.Unlock()
-
+	updateRulesMutex.Lock()
+	defer updateRulesMutex.Unlock()
 	for _, ports := range spec.Ports {
 		if err = spec.securityListManager.Update(ctx, lbSubnets, nodeSubnets, spec.SourceCIDRs, nil, ports, *spec.IsPreserveSource); err != nil {
 			return err
@@ -898,18 +997,17 @@ func (clb *CloudLoadBalancerProvider) updateBackendSet(ctx context.Context, lbID
 		"loadBalancerID", lbID,
 		"loadBalancerType", getLoadBalancerType(spec.service))
 	logger.Info("Applying action on backend set")
-
 	switch action.Type() {
 	case Create:
 		err = secListManager.Update(ctx, lbSubnets, nodeSubnets, sourceCIDRs, nil, ports, *spec.IsPreserveSource)
 		if err != nil {
 			return err
 		}
-
 		workRequestID, err = clb.lbClient.CreateBackendSet(ctx, lbID, action.Name(), &bs)
 	case Update:
 		// For NLB, due to source IP preservation we need to ensure ingress rules from sourceCIDRs are added to
 		// the backends subnet's seclist as well
+
 		if err = secListManager.Update(ctx, lbSubnets, nodeSubnets, spec.SourceCIDRs, action.OldPorts, ports, *spec.IsPreserveSource); err != nil {
 			return err
 		}
@@ -919,7 +1017,6 @@ func (clb *CloudLoadBalancerProvider) updateBackendSet(ctx context.Context, lbID
 		if err != nil {
 			return err
 		}
-
 		workRequestID, err = clb.lbClient.DeleteBackendSet(ctx, lbID, action.Name())
 	}
 
@@ -950,28 +1047,24 @@ func (clb *CloudLoadBalancerProvider) updateListener(ctx context.Context, lbID s
 		"loadBalancerID", lbID,
 		"loadBalancerType", getLoadBalancerType(spec.service))
 	logger.Info("Applying action on listener")
-
 	switch action.Type() {
 	case Create:
 		err = secListManager.Update(ctx, lbSubnets, nodeSubnets, sourceCIDRs, nil, ports, *spec.IsPreserveSource)
 		if err != nil {
 			return err
 		}
-
 		workRequestID, err = clb.lbClient.CreateListener(ctx, lbID, action.Name(), &listener)
 	case Update:
 		err = secListManager.Update(ctx, lbSubnets, nodeSubnets, sourceCIDRs, nil, ports, *spec.IsPreserveSource)
 		if err != nil {
 			return err
 		}
-
 		workRequestID, err = clb.lbClient.UpdateListener(ctx, lbID, action.Name(), &listener)
 	case Delete:
 		err = secListManager.Delete(ctx, lbSubnets, nodeSubnets, ports, sourceCIDRs, *spec.IsPreserveSource)
 		if err != nil {
 			return err
 		}
-
 		workRequestID, err = clb.lbClient.DeleteListener(ctx, lbID, action.Name())
 	}
 
@@ -1043,6 +1136,7 @@ func (cp *CloudProvider) EnsureLoadBalancerDeleted(ctx context.Context, clusterN
 
 	var errorType string
 	var lbMetricDimension string
+	var nsgMetricDimension string
 
 	dimensionsMap := make(map[string]string)
 
@@ -1068,14 +1162,37 @@ func (cp *CloudProvider) EnsureLoadBalancerDeleted(ctx context.Context, clusterN
 	id := *lb.Id
 	dimensionsMap[metrics.ResourceOCIDDimension] = id
 	logger = logger.With("loadBalancerID", id, "loadBalancerType", getLoadBalancerType(service))
-	secListManagerMode, err := getSecurityListManagementMode(service)
+	securityRuleManagementMode, nsg, err := getRuleManagementMode(service)
 	if err != nil {
-		logger.With(zap.Error(err)).Error("failed to get seclist management mode")
-		return errors.Wrap(err, "failed to get seclist management mode")
+		logger.With(zap.Error(err)).Error("failed to get rule management mode")
+		return errors.Wrap(err, "failed to get rule management mode")
 	}
+
+	var etag *string
+	var frontendNsgId = ""
+	if securityRuleManagementMode == NSG {
+		// List network security groups
+		nsgs := lb.NetworkSecurityGroupIds
+		uid := fmt.Sprintf("%s", service.UID)
+		for _, nsgId := range nsgs {
+			frontendNsgId, etag, err = cp.getFrontendNsg(ctx, logger, nsgId, uid)
+			if err != nil {
+				errorType = util.GetError(err)
+				nsgMetricDimension = util.GetMetricDimensionForComponent(errorType, util.NSGType)
+				dimensionsMap[metrics.ComponentDimension] = nsgMetricDimension
+				metrics.SendMetricData(cp.metricPusher, getMetric(util.NSGType, Get), time.Since(startTime).Seconds(), dimensionsMap)
+			}
+			if frontendNsgId != "" {
+				logger.With("frontendNsgId", frontendNsgId).Infof("queue frontend nsg for delete %s", frontendNsgId)
+				nsg.frontendNsgId = frontendNsgId
+				break
+			}
+		}
+	}
+
 	// get annotation from load balancer spec and compare to ManagementModeNone
-	if secListManagerMode != ManagementModeNone {
-		err := cp.cleanupSecListForLoadBalancerDelete(lb, logger, ctx, service, name)
+	if securityRuleManagementMode != ManagementModeNone {
+		err := cp.cleanupSecurityRulesForLoadBalancerDelete(lb, logger, ctx, service, name, frontendNsgId)
 		if err != nil {
 			errorType = util.GetError(err)
 			lbMetricDimension = util.GetMetricDimensionForComponent(errorType, util.LoadBalancerType)
@@ -1085,6 +1202,7 @@ func (cp *CloudProvider) EnsureLoadBalancerDeleted(ctx context.Context, clusterN
 			return err
 		}
 	}
+
 	logger.Info("Deleting load balancer")
 	workReqID, err := lbProvider.lbClient.DeleteLoadBalancer(ctx, id)
 	if err != nil {
@@ -1111,22 +1229,42 @@ func (cp *CloudProvider) EnsureLoadBalancerDeleted(ctx context.Context, clusterN
 	lbMetricDimension = util.GetMetricDimensionForComponent(util.Success, util.LoadBalancerType)
 	dimensionsMap[metrics.ComponentDimension] = lbMetricDimension
 	metrics.SendMetricData(cp.metricPusher, getMetric(loadBalancerType, Delete), time.Since(startTime).Seconds(), dimensionsMap)
+
+	// Delete of NSG happens after delete of the Loadbalancer
+	if nsg != nil && nsg.nsgRuleManagementMode == RuleManagementModeNsg && frontendNsgId != "" {
+		if etag != nil {
+			nsgDeleted, err := cp.deleteNsg(ctx, logger, nsg.frontendNsgId, *etag)
+			if !nsgDeleted || err != nil {
+				logger.With(zap.Error(err)).Error("failed to delete nsg")
+				errorType = util.GetError(err)
+				nsgMetricDimension = util.GetMetricDimensionForComponent(errorType, util.NSGType)
+				dimensionsMap[metrics.ComponentDimension] = nsgMetricDimension
+				metrics.SendMetricData(cp.metricPusher, getMetric(util.NSGType, Delete), time.Since(startTime).Seconds(), dimensionsMap)
+				return err
+			}
+			nsgMetricDimension = util.GetMetricDimensionForComponent(util.Success, util.NSGType)
+			dimensionsMap[metrics.ComponentDimension] = nsgMetricDimension
+			metrics.SendMetricData(cp.metricPusher, getMetric(util.NSGType, Delete), time.Since(startTime).Seconds(), dimensionsMap)
+			logger.With("frontendNsgId", nsg.frontendNsgId).Infof("ccm managed nsg with id %s deleted", nsg.frontendNsgId)
+		}
+	}
+
 	return nil
 }
 
 // Critical Section for Security List Updates
-func (cp *CloudProvider) cleanupSecListForLoadBalancerDelete(lb *client.GenericLoadBalancer, logger *zap.SugaredLogger, ctx context.Context, service *v1.Service, name string) error {
-	securityListMutex.Lock()
-	defer securityListMutex.Unlock()
+func (cp *CloudProvider) cleanupSecurityRulesForLoadBalancerDelete(lb *client.GenericLoadBalancer, logger *zap.SugaredLogger, ctx context.Context, service *v1.Service, name string, frontendNsgOcid string) error {
+	updateRulesMutex.Lock()
+	defer updateRulesMutex.Unlock()
 
 	id := *lb.Id
-	nodeIPs := sets.NewString()
+	ipSet := sets.NewString()
 	for _, backendSet := range lb.BackendSets {
 		for _, backend := range backendSet.Backends {
-			nodeIPs.Insert(*backend.IpAddress)
+			ipSet.Insert(*backend.IpAddress)
 		}
 	}
-	nodes, err := cp.getNodesByIPs(nodeIPs.List())
+	nodes, err := cp.getNodesByIPs(ipSet.List())
 	if err != nil {
 		logger.With(zap.Error(err)).Error("Failed to fetch nodes by internal ips")
 		return errors.Wrap(err, "fetching nodes by internal ips")
@@ -1142,19 +1280,57 @@ func (cp *CloudProvider) cleanupSecListForLoadBalancerDelete(lb *client.GenericL
 		logger.With(zap.Error(err)).Error("Failed to get subnets for load balancers")
 		return errors.Wrap(err, "getting subnets for load balancers")
 	}
-	secListManagerMode, err := getSecurityListManagementMode(service)
+
+	securityRuleManagerMode, managedNsg, err := getRuleManagementMode(service)
 	if err != nil {
-		logger.With(zap.Error(err)).Error("failed to get seclist management mode")
-		return errors.Wrap(err, "failed to get seclist management mode")
+		logger.With(zap.Error(err)).Error("failed to get Security Rule management mode")
+		return errors.Wrap(err, "failed to get Security Rule management mode")
 	}
 
-	securityListManager := cp.securityListManagerFactory(
-		secListManagerMode)
+	backendNsgIds, err := getManagedBackendNSG(service)
+	if err != nil {
+		logger.With(zap.Error(err)).Error("failed to get backend Nsgs from spec")
+		return errors.Wrap(err, "failed to get backend Nsgs from spec")
+	}
+	var securityListManager securityListManager
+	if securityRuleManagerMode == NSG {
+		if managedNsg != nil {
+			if frontendNsgOcid != "" {
+				managedNsg.frontendNsgId = frontendNsgOcid
+			}
+			if len(backendNsgIds) > 0 {
+				managedNsg.backendNsgId = backendNsgIds
+			}
+		}
+	} else {
+		securityListManager = cp.securityListManagerFactory(
+			securityRuleManagerMode)
+	}
 
 	isPreserveSource, err := getPreserveSource(logger, service)
 	if err != nil {
 		logger.With(zap.Error(err)).Error("failed to determine value for is-preserve-source")
 		return errors.Wrap(err, "failed to determine value for is-preserve-source")
+	}
+	portsNsg, err := getPorts(service)
+	if err != nil {
+		return errors.Wrapf(err, "failed to get ports from spec")
+	}
+	sourceCIDRs, err := getLoadBalancerSourceRanges(service)
+	if securityRuleManagerMode == NSG && len(managedNsg.backendNsgId) > 0 {
+		serviceComponents := serviceComponents{
+			frontendNsgOcid:  managedNsg.frontendNsgId,
+			backendNsgOcids:  managedNsg.backendNsgId,
+			sourceCIDRs:      sourceCIDRs,
+			ports:            portsNsg,
+			isPreserveSource: isPreserveSource,
+			serviceUid:       fmt.Sprintf("service-uid-%s", service.UID),
+		}
+		logger.Infof("(ensureloadbalancer deleted) Service Components %#v", serviceComponents)
+		err = cp.removeBackendSecurityGroupRules(ctx, serviceComponents)
+		if err != nil {
+			return err
+		}
 	}
 
 	for listenerName, listener := range lb.Listeners {
@@ -1176,10 +1352,13 @@ func (cp *CloudProvider) cleanupSecListForLoadBalancerDelete(lb *client.GenericL
 			return errors.Wrapf(err, "delete security rules for listener %q on load balancer %q", listenerName, name)
 		}
 
-		if err = securityListManager.Delete(ctx, lbSubnets, nodeSubnets, ports, sourceCIDRs, isPreserveSource); err != nil {
-			logger.With(zap.Error(err)).Errorf("Failed to delete security rules for listener %q on load balancer %q", listenerName, name)
-			return errors.Wrapf(err, "delete security rules for listener %q on load balancer %q", listenerName, name)
+		if securityRuleManagerMode == ManagementModeAll || securityRuleManagerMode == ManagementModeFrontend {
+			if err = securityListManager.Delete(ctx, lbSubnets, nodeSubnets, ports, sourceCIDRs, isPreserveSource); err != nil {
+				logger.With(zap.Error(err)).Errorf("Failed to delete security rules for listener %q on load balancer %q", listenerName, name)
+				return errors.Wrapf(err, "delete security rules for listener %q on load balancer %q", listenerName, name)
+			}
 		}
+
 	}
 	return nil
 }
@@ -1265,4 +1444,32 @@ func (cp *CloudProvider) checkAllBackendNodesNotReady(nodeList []*v1.Node) bool 
 		}
 	}
 	return true
+}
+
+// If CCM manages the NSG for the service, CCM to delete the NSG when the LB/NLB service is deleted
+func (cp *CloudProvider) deleteNsg(ctx context.Context, logger *zap.SugaredLogger, id, etag string) (bool, error) {
+	opcRequestId, err := cp.client.Networking().DeleteNetworkSecurityGroup(ctx, id, etag)
+	if err != nil {
+		logger.Errorf("failed to delete nsg %s OpcRequestId %s", id, *opcRequestId)
+		return false, err
+	}
+	logger.Infof("delete nsg OpcRequestId %s", *opcRequestId)
+	return true, nil
+}
+
+func (cp *CloudProvider) getFrontendNsg(ctx context.Context, logger *zap.SugaredLogger, id, uid string) (string, *string, error) {
+	nsg, etag, err := cp.client.Networking().GetNetworkSecurityGroup(ctx, id)
+	if err != nil || etag == nil {
+		logger.Errorf("failed to get nsg %s", id)
+		return "", nil, err
+	}
+	logger.Infof("Get nsg %s", *nsg)
+	freeFormTags := map[string]string{"CreatedBy": "OKE-CCM", "ServiceUid": uid}
+	if reflect.DeepEqual(nsg.FreeformTags, freeFormTags) {
+		logger.Infof("Found OKE created frontend nsg %s", *nsg.Id)
+		return *nsg.Id, etag, nil
+	} else {
+		logger.Infof("Found attached nsg %s but not OKE created", *nsg.Id)
+		return "", nil, nil
+	}
 }
